--- apps/nextjs/src/components/features/inbox/index.ts
@@ -1,9 +1,0 @@
-export * from './hooks';
-export * from './inbox';
-export * from './inbox-header';
-export * from './inbox-left';
-export * from './inbox-list';
-export * from './inbox-list-content';
-export * from './inbox-list-header';
-export * from './inbox-right';
-export * from './inbox-skeleton';
